refactor: jobs module
