+++ src/modules/users/users.module.ts
@@ -4,9 +4,10 @@ import { User } from './entities/user.entity';
 import { UserRepository } from './repositories/user.repository';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
+import { JobsModule } from '../jobs/jobs.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User])],
+  imports: [TypeOrmModule.forFeature([User]), JobsModule],
   providers: [UsersService, UserRepository],
   controllers: [UsersController],
 })
